Changes of version: 20070722

+++ libewf/ewf_compress.c
@@ -49,7 +49,7 @@
 /* Compresses data, wraps zlib uncompress function
  * Returns 1 on success, -1 on error
  */
-int8_t ewf_compress( uint8_t *compressed_data, size_t *compressed_size, uint8_t *uncompressed_data, size_t uncompressed_size, int8_t compression_level )
+int ewf_compress( uint8_t *compressed_data, size_t *compressed_size, uint8_t *uncompressed_data, size_t uncompressed_size, int8_t compression_level )
 {
 	static char *function        = "ewf_compress";
 	uLongf safe_compressed_size  = 0;
@@ -158,7 +158,7 @@ int8_t ewf_compress( uint8_t *compressed_data, size_t *compressed_size, uint8_t 
 /* Uncompresses data, wraps zlib uncompress function
  * Returns 1 on success, 0 on failure, -1 on error
  */
-int8_t ewf_uncompress( uint8_t *uncompressed_data, size_t *uncompressed_size, uint8_t *compressed_data, size_t compressed_size )
+int ewf_uncompress( uint8_t *uncompressed_data, size_t *uncompressed_size, uint8_t *compressed_data, size_t compressed_size )
 {
 	static char *function         = "ewf_uncompress";
 	uLongf safe_uncompressed_size = 0;
